requirements-dev.txt
# Development-only tools; includes runtime too
-r requirements.txt

pre-commit
black
ruff
pytest
pytest-django
django-browser-reload
django-extensions
